utils: refactor generate_output_file in sct.image
